requirements-py2.txt
azure-identity==1.7.1
